+++ docs/examples/crawl_all_links.mdx
@@ -17,7 +17,7 @@ as the crawler navigates from page to page. This example can also be used to fin
 
 :::tip
 
-If no options are given, by default the method will only add links that are under the same subdomain. This behavior can be controlled with the <ApiLink to="core/interface/EnqueueUrlsOptions#strategy">`strategy`</ApiLink>
+If no options are given, by default the method will only add links that are under the same subdomain. This behavior can be controlled with the <ApiLink to="basic-crawler/interface/EnqueueUrlsOptions#strategy">`strategy`</ApiLink>
 option. You can find more info about this option in the [`Crawl relative links`](./crawl-relative-links) examples.
 
 :::
