+++ docs/introduction/03-adding-urls.mdx
@@ -130,7 +130,7 @@ await enqueueLinks({
 
 ### Filter URLs with patterns
 
-For even more control, you can use `include` and `exclude` to filter the URLs. Each accepts an `Array` of glob pattern strings, `{ glob: string }` objects, `RegExp` instances, or `{ regexp: RegExp }` objects. <ApiLink to="core/interface/EnqueueUrlsOptions">See the reference</ApiLink> for more information about them as well as other options.
+For even more control, you can use `include` and `exclude` to filter the URLs. Each accepts an `Array` of glob pattern strings, `{ glob: string }` objects, `RegExp` instances, or `{ regexp: RegExp }` objects. <ApiLink to="basic-crawler/interface/EnqueueUrlsOptions">See the reference</ApiLink> for more information about them as well as other options.
 
 :::caution Defaults override
 
@@ -146,7 +146,7 @@ await enqueueLinks({
 
 ### Transform requests
 
-To have absolute control, we have the <ApiLink to="core/interface/EnqueueUrlsOptions#transformRequestFunction">`transformRequestFunction`</ApiLink>. After request options are filtered by `include`/`exclude` patterns, this function can be used to skip them or modify their contents such as `userData`, `payload` or, most importantly, `uniqueKey`. This is useful when you need to enqueue multiple requests to the queue, and these requests share the same URL, but differ in methods or payloads. Another use case is to dynamically update or create the `userData`.
+To have absolute control, we have the <ApiLink to="basic-crawler/interface/EnqueueUrlsOptions#transformRequestFunction">`transformRequestFunction`</ApiLink>. After request options are filtered by `include`/`exclude` patterns, this function can be used to skip them or modify their contents such as `userData`, `payload` or, most importantly, `uniqueKey`. This is useful when you need to enqueue multiple requests to the queue, and these requests share the same URL, but differ in methods or payloads. Another use case is to dynamically update or create the `userData`.
 
 ```ts
 await enqueueLinks({
