+++ packages/basic-crawler/src/index.ts
@@ -3,6 +3,7 @@ export * from './internals/autoscaling/index.js';
 export * from './internals/basic-crawler.js';
 export * from './internals/cookie_utils.js';
 export * from './internals/crawlers/index.js';
+export * from './internals/enqueue_links/index.js';
 export * from './internals/errors.js';
 export * from './internals/http.js';
 export * from './internals/router.js';
