+++ packages/core/package.json
@@ -64,9 +64,7 @@
         "csv-stringify": "^6.5.2",
         "json5": "^2.2.3",
         "mime-types": "^3.0.1",
-        "minimatch": "^10.0.1",
         "stream-json": "^1.9.1",
-        "tldts": "^7.0.6",
         "tslib": "^2.8.1",
         "type-fest": "^4.41.0",
         "zod": "catalog:"
