+++ packages/core/src/index.ts
@@ -2,7 +2,6 @@ export * from './debug.js';
 export * from './errors.js';
 export * from './configuration.js';
 export * from './service_locator.js';
-export * from './enqueue_links/index.js';
 export * from './events/index.js';
 export * from './log.js';
 export * from './owned_or_injected.js';
@@ -16,3 +15,4 @@ export * from './memory-storage/index.js';
 export { ArgumentValidationError, validators } from './validators.js';
 export * from './recoverable_state.js';
 export type { StorageBackend } from '@crawlee/types';
+export { EnqueueStrategy } from '@crawlee/utils';
