+++ packages/core/src/request.ts
@@ -3,15 +3,26 @@ import crypto from 'node:crypto';
 import util from 'node:util';
 
 import type { AllowedHttpMethods, Dictionary } from '@crawlee/types';
+import type { EnqueueStrategy } from '@crawlee/utils';
 import { z } from 'zod';
 
 import { cryptoRandomObjectId, normalizeUrl } from '@apify/utilities';
 
-import type { EnqueueStrategyOption } from './enqueue_links/enqueue_links.js';
-import type { SkippedRequestReason } from './enqueue_links/shared.js';
 import { serviceLocator } from './service_locator.js';
 import { keys } from './typedefs.js';
 import { parseArgument, schemas } from './validators.js';
+
+/** The `strategy` option accepted by {@apilink ExtractLinksOptions} and {@apilink EnqueueUrlsOptions}. */
+export type EnqueueStrategyOption = EnqueueStrategy | 'all' | 'same-domain' | 'same-hostname' | 'same-origin';
+
+export type SkippedRequestReason =
+    | 'robotsTxt'
+    | 'limit'
+    | 'enqueueLimit'
+    | 'filters'
+    | 'transform'
+    | 'redirect'
+    | 'depth';
 
 const dateString = z.string().refine((value) => !Number.isNaN(Date.parse(value)), {
     message: 'Invalid input: expected a date string',
